source/python-requirements.txt
aenum==2.2.6
bottle==0.12.19
bottle-websocket==0.2.9
canopen==1.2.0
cffi==1.14.4
Eel==0.14.0
Events==0.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
pycparser==2.20
pyparsing==2.4.7
pyserial==3.5
python-can==3.3.4
PyTrinamic==0.1.27
whichcraft==0.6.1
windows-curses==2.2.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0